refactoring: delete entity to service
