+++ application/src/main/java/org/thingsboard/server/controller/DeviceProfileController.java
@@ -71,8 +71,7 @@ import static org.thingsboard.server.controller.ControllerConstants.UUID_WIKI_LI
 @TbCoreComponent
 @RequestMapping("/api")
 @Slf4j
-public class DeviceProfileController extends BaseController {
-
+public class DeviceProfileController extends DefaultEntityBaseController {
 
     @Autowired
     private TimeseriesService timeseriesService;
@@ -252,26 +251,7 @@ public class DeviceProfileController extends BaseController {
             @ApiParam(value = DEVICE_PROFILE_ID_PARAM_DESCRIPTION)
             @PathVariable(DEVICE_PROFILE_ID) String strDeviceProfileId) throws ThingsboardException {
         checkParameter(DEVICE_PROFILE_ID, strDeviceProfileId);
-        try {
-            DeviceProfileId deviceProfileId = new DeviceProfileId(toUUID(strDeviceProfileId));
-            DeviceProfile deviceProfile = checkDeviceProfileId(deviceProfileId, Operation.DELETE);
-            deviceProfileService.deleteDeviceProfile(getTenantId(), deviceProfileId);
-
-            tbClusterService.onDeviceProfileDelete(deviceProfile, null);
-            tbClusterService.broadcastEntityStateChangeEvent(deviceProfile.getTenantId(), deviceProfile.getId(), ComponentLifecycleEvent.DELETED);
-
-            logEntityAction(deviceProfileId, deviceProfile,
-                    null,
-                    ActionType.DELETED, null, strDeviceProfileId);
-
-            sendEntityNotificationMsg(getTenantId(), deviceProfile.getId(), EdgeEventActionType.DELETED);
-        } catch (Exception e) {
-            logEntityAction(emptyId(EntityType.DEVICE_PROFILE),
-                    null,
-                    null,
-                    ActionType.DELETED, e, strDeviceProfileId);
-            throw handleException(e);
-        }
+        entityDeleteService.deleteEntity(getTenantId(), new DeviceProfileId(toUUID(strDeviceProfileId)));
     }
 
     @ApiOperation(value = "Make Device Profile Default (setDefaultDeviceProfile)",
